+++ application/controllers/Charlie.php
@@ -1,7 +1,7 @@
 <?php
 defined('BASEPATH') OR exit('No direct script access allowed');
 
-class Able extends Application
+class Charlie extends Application
 {
 
 	function __construct()
@@ -11,11 +11,11 @@ class Able extends Application
 
 	/**
 	 * Alpha Regular Controller
-         * Small Changed
+     * Small Changed
 	 */
-	public function index()
+	public function brown()
 	{
-		$this->show(1);
+		$this->show(3);
 	}
 
 }