+++ changelog.md
@@ -1,6 +1,12 @@
 Harshita - Captain
 Takito (Hiram) - Programming Minion
 
-Version 1.0
+Version 3.0
 
 Log Steps:
+
+-> Completed Able Issue #1 ~Takito
+-> Completed Bravo Issue #2 ~Harshita
+-> Completed Charlie Issue #3 ~Takito
+-> Completed Delta Issue #4 ~Harshita
+-> Completed Foxtrot Issue #6 ~Harshita